상하단 만들기
