+++ src/App.jsx
@@ -1,19 +1,20 @@
-//import React from "react";
-import { Routes, Route } from 'react-router-dom';
-
-//import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
-import MyPage from "./routes/MyPage";
+import React from "react";
+import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import "./App.css";
-import profileIcon from "./assets/icons/profileIcon.png";
+import profileIcon from "./assets/icons/Header/profileIcon.png";
+import LayOut from './components/commmon/LayOut';
+import MyPage from "./pages/MyPage";
 import MainPage from './pages/MainPage';
 
 function App() {
   return (
     <Routes>
-        <Route>
-          <Route path={"/MyPage"} element={<MyPage profileIcon={profileIcon} />} />
-          <Route path={"/MainPage"} element={<MainPage/>}/>
+      <Route>
+      <Route element={<LayOut />}>
+          <Route path="/MyPage" element={<MyPage profileIcon={profileIcon} />} />
+          <Route path="/MainPage" element={<MainPage />} />
         </Route>
+      </Route>
     </Routes>
   );
 }
