+++ src/components/commmon/Footer.jsx
@@ -1,23 +1,42 @@
 import React from 'react';
 import styled from 'styled-components';
 
+import Logo from "../../assets/icons/TemplateCard/templateThumnail.png";
+import Facebook from "../../assets/icons/Footer/Facebook.png";
+import Instagram from "../../assets/icons/Footer/Instagram.png";
+import Twitter from "../../assets/icons/Footer/Twitter.png";
+
 const Footer = () => {
     return(
         <FooterWrapper>
             <MainText>FolioFrame</MainText>
-            {/*<LogoWrapper>*/}
-            <rightTextWrapper>
-                <Text>팀프로젝트</Text>
-            </rightTextWrapper>
-               
+            <LogoWrapper>
+                {/*로고 대신해서...;;*/}
+                <img src={Logo} alt="Logo" width={50} height={50}/>
+            </LogoWrapper>
+
+            <TextWrapper>
+                <RightWrapper>
+                    <Text>팀프로젝트1 5팀</Text>
+                    <Text>팀명 : 포폴만드조</Text>
+                    <Text>조원 : 김태연, 김예은, 조수연, 최현혜</Text>
+                </RightWrapper>
+                
+                <LeftWrapper>
+                    <img src={Twitter} alt="Twitter" /> 
+                    <img src={Instagram} alt="Instagram" />
+                    <img src={Facebook} alt="Facebook" />
+                </LeftWrapper>
+            </TextWrapper>
         </FooterWrapper>
     );
 };
 
 export default Footer;
 
+// css Wrapper
 const FooterWrapper = styled.div`
-    background-coloer : #D0D1D9;
+    background-color : #D0D1D9;
     width : 100%;
     height: 277px;
     box-sizing: border-box;
@@ -25,22 +44,42 @@ const FooterWrapper = styled.div`
     flex-direction: column;
     align-items: center;
     justify-content: center;
+    padding : 40px 40px;  /*margin값 40으로 주기로 해서 padding에 위아래 40으로 쥼!!!*/
 `;
 
+const LogoWrapper = styled.div`
+    align-items: center;
+    justify-content: center;
+`;
+
+const TextWrapper = styled.div`
+    display: flex;
+    justify-content: space-between; 
+    align-items: center; 
+    width: 100%;
+`;
+
+const RightWrapper = styled.div`
+    color: #fff;
+`;
+
+const LeftWrapper = styled.div`
+    display: flex;
+    gap: 10px; 
+ `;
+
+// css Text
 const MainText = styled.div`
-    font-size: 1em;
+    font-size: 1.5em;
     font-weight: 700;
     text-align: center;
-`;
-
-const rightTextWrapper = styled.div`
-
+    color: #fff;
 `;
 
 const Text = styled.div`
-    color: #697077;
-    font-size: 0.8125em;
-    font-weight: 400;
+    color: #fff;
+    font-size: 0.9em;
+    font-weight: 600;
     cursor: pointer;
-
+    padding: 2px; 0;
 `;
