+++ src/components/commmon/LayOut.jsx
@@ -1,0 +1,23 @@
+import React from 'react';
+import { Outlet } from 'react-router-dom';
+import Header from './Header.jsx';
+import Footer from './footer.jsx';
+import styled from 'styled-components';
+
+const LayOut = () => {
+    return (
+        <>
+            <Header/>
+            <Padding>
+                <Outlet/>
+            </Padding>
+            <Footer/>
+        </>
+    );
+};
+
+export default LayOut;
+
+const Padding = styled.div`
+    padding-top: 60px;
+`;
