+++ src/pages/MainPage.jsx
@@ -1,11 +1,12 @@
 import React from "react";
 import Footer from "../components/commmon/footer";
+import Header from "../components/commmon/Header";
 const MainPage = () => {
     return(
         <>
-        
+
         <p>ddd</p>
-        <Footer/>
+   
         </>
     );
 };
